requirements.txt
TgCrypto
Pyrogram
requests
python-dotenv